deleted:    components/chatbox.inc.php 	deleted:    components/economyoverview.inc.php 	deleted:    components/head.inc.php 	deleted:    components/infobox.inc.php 	deleted:    components/serverstatus.inc.php 	deleted:    components/todo.inc.php 	deleted:    components/userpanel.inc.php 	modified:   css/style_register.css 	renamed:    db_config.php -> database/db_config.php 	modified:   delete_task.php 	modified:   fetch_messages.php 	renamed:    chat_functions.php -> functions/chat_functions.php 	renamed:    gather_info.php -> functions/gather_info.php 	modified:   get_messages.php 	modified:   get_most_present_username.php 	modified:   get_online_users.php 	modified:   index.php 	modified:   join_or_leave_chat.php 	new file:   js/register.js 	modified:   login.php 	modified:   register.php 	modified:   save_message.php

+++ delete_task.php
@@ -1,14 +1,20 @@
 <?php
-include 'db_config.php';
+$servername = 'localhost';
+$db_username = 'root';
+$password = '';
+$database = 'webdev';
 
-$conn = connect_webserver_db();
+$conn = new mysqli($servername, $db_username, $password, $database);
+if ($conn->connect_error) {
+    die('Connection failed: ' . $conn->connect_error);
+}
 
 if (isset($_POST['taskId'])) {
     $taskId = $_POST['taskId'];
 
     $sql = 'DELETE FROM todo_tasks WHERE id = ?';
     $stmt = $conn->prepare($sql);
-    $stmt->bindParam(1, $taskId, PDO::PARAM_INT);
+    $stmt->bind_param('i', $taskId);
     if ($stmt->execute()) {
         $response = ['success' => true];
         echo json_encode($response);
@@ -21,6 +27,6 @@ if (isset($_POST['taskId'])) {
     echo json_encode($response);
 }
 
-$stmt = null;
-$conn = null;
+$stmt->close();
+$conn->close();
 ?>
